refactor: rename plugin module to scm_provider to avoid conflict
Main branch added src/plugin.rs for MCP plugin management. Our SCM
provider module at src/plugin/ conflicted (Rust E0761: both plugin.rs
and plugin/mod.rs found). Renamed to src/scm_provider/ and updated
all references. Also renamed list_plugins command to list_scm_providers
to avoid collision with main's plugin commands.
