+++ src/scm_provider/manifest.rs
