+++ src/ui/src/services/tauri.ts
@@ -737,6 +737,7 @@ export function openUrl(url: string): Promise<void> {
   return invoke("open_url", { url });
 }
 
+
 export function listScmProviders(): Promise<PluginInfo[]> {
   return invoke("list_scm_providers");
 }
